Requiring objects which wish to store their state to inherit from a class was originally intended to facilitate automatic persistence. The engine would persist all properties of the class instance when appropriate without scripter direction. There turn out to be no smart ways to automatically determine how to store array/dictionary contents, however, as they may contain references to other objects, and there is no easy, reliable way in AS to tell whether a reference is owning or weak. It would be much better to ask the scripter than to try to guess.
The cleaner approach is to provide an ObjectState class which stores properties
and values in a table/a-list and which may be used with any object, and that
buffers writes to the database. Aside from avoiding pollution of every
persistent object's class hierarchy, this would also make it natural to store or
apply previous states of an object.